Add option to include pending txn count in nonce (#943)
* Add option to include pending txn count in nonce

- `CAPE_RELAYER_NONCE_COUNT_RULE` can be set to `mined` or `pending`.
- Add CLI argument for `CAPE_RELAYER_PORT`.
- Change some port variables to u16 type.
- Run `test_relay` with both nonce count rules.

+++ wallet/src/testing.rs
@@ -57,13 +57,13 @@ use tempdir::TempDir;
 use tracing::{event, Level};
 
 lazy_static! {
-    static ref PORT: Arc<Mutex<u64>> = {
+    static ref PORT: Arc<Mutex<u16>> = {
         let port_offset = std::env::var("PORT").unwrap_or_else(|_| String::from("60000"));
         Arc::new(Mutex::new(port_offset.parse().unwrap()))
     };
 }
 
-pub async fn port() -> u64 {
+pub async fn port() -> u16 {
     let mut counter = PORT.lock().await;
     let port = *counter;
     *counter += 1;
